pyproject.toml
[project]
name = "librarian"
version = "0.1.0"
description = "A vibe-based book discovery engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "strands-agents[gemini]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "langdetect>=1.0.9",
    "exa-py>=1.0.0",
    "tavily-python>=0.3.0",
    "colorama>=0.4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
